backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
yt-dlp==2023.11.16
python-socketio==5.11.1 